Fixed permission issue on project setup
In before, we setup the project by running sudo on normal user,
which ran into issue of permission. In addition, for some account,
the shell prompt for password on sudo which prevent the script to
continue. The only way to prevent password prompting is to modify
the sudoer setting, which is very risky thing to do. Therefore,
I changed the way to setup the app in this commit by requiring
the script is executed under root(non-root user will get permission
denied) and su to the user who own the applicaiton.